+++ connect/lib/core/src/main/scala/io/iohk/atala/connect/core/repository/ConnectionRepository.scala
@@ -23,10 +23,29 @@ trait ConnectionRepository[F[_]] {
 
   def getConnectionRecordByThreadId(thid: UUID): F[Option[ConnectionRecord]]
 
-  def updateWithConnectionRequest(recordId: UUID, request: ConnectionRequest, state: ProtocolState): F[Int]
+  def updateWithConnectionRequest(
+      recordId: UUID,
+      request: ConnectionRequest,
+      state: ProtocolState,
+      maxRetries: Int, // max numbre of retries -> set the metaRetries
+  ): F[Int]
 
-  def updateWithConnectionResponse(recordId: UUID, response: ConnectionResponse, state: ProtocolState): F[Int]
+  def updateWithConnectionResponse(
+      recordId: UUID,
+      response: ConnectionResponse,
+      state: ProtocolState,
+      maxRetries: Int, // max numbre of retries -> set the metaRetries
+  ): F[Int]
 
-  def updateConnectionProtocolState(recordId: UUID, from: ProtocolState, to: ProtocolState): F[Int]
+  def updateConnectionProtocolState(
+      recordId: UUID,
+      from: ProtocolState,
+      to: ProtocolState,
+      maxRetries: Int, // max numbre of retries -> set the metaRetries
+  ): F[Int]
 
+  def updateAfterFail(
+      recordId: UUID,
+      failReason: Option[String],
+  ): F[Int]
 }