+++ connect/lib/core/src/main/scala/io/iohk/atala/connect/core/service/ConnectionServiceImpl.scala
@@ -17,7 +17,8 @@ import io.iohk.atala.mercury.protocol.connection.ConnectionResponse
 import io.iohk.atala.shared.utils.Base64Utils
 
 private class ConnectionServiceImpl(
-    connectionRepository: ConnectionRepository[Task]
+    connectionRepository: ConnectionRepository[Task],
+    maxRetries: Int = 5, // TODO move to config
 ) extends ConnectionService {
 
   override def createConnectionInvitation(
@@ -37,7 +38,9 @@ private class ConnectionServiceImpl(
           protocolState = ConnectionRecord.ProtocolState.InvitationGenerated,
           invitation = invitation,
           connectionRequest = None,
-          connectionResponse = None
+          connectionResponse = None,
+          metaRetries = maxRetries,
+          metaLastFailure = None,
         )
       )
       count <- connectionRepository
@@ -102,7 +105,9 @@ private class ConnectionServiceImpl(
           protocolState = ConnectionRecord.ProtocolState.InvitationReceived,
           invitation = invitation,
           connectionRequest = None,
-          connectionResponse = None
+          connectionResponse = None,
+          metaRetries = maxRetries,
+          metaLastFailure = None,
         )
       )
       count <- connectionRepository
@@ -122,7 +127,7 @@ private class ConnectionServiceImpl(
       record <- getRecordWithState(recordId, ProtocolState.InvitationReceived)
       request = createDidCommConnectionRequest(record, pairwiseDid)
       count <- connectionRepository
-        .updateWithConnectionRequest(recordId, request, ProtocolState.ConnectionRequestPending)
+        .updateWithConnectionRequest(recordId, request, ProtocolState.ConnectionRequestPending, maxRetries)
         .mapError(RepositoryError.apply)
       _ <- count match
         case 1 => ZIO.succeed(())
@@ -145,7 +150,7 @@ private class ConnectionServiceImpl(
     for {
       record <- getRecordFromThreadIdAndState(request.thid, ProtocolState.InvitationGenerated)
       _ <- connectionRepository
-        .updateWithConnectionRequest(record.id, request, ProtocolState.ConnectionRequestReceived)
+        .updateWithConnectionRequest(record.id, request, ProtocolState.ConnectionRequestReceived, maxRetries)
         .flatMap {
           case 1 => ZIO.succeed(())
           case n => ZIO.fail(UnexpectedException(s"Invalid row count result: $n"))
@@ -161,7 +166,7 @@ private class ConnectionServiceImpl(
       record <- getRecordWithState(recordId, ProtocolState.ConnectionRequestReceived)
       response = createDidCommConnectionResponse(record)
       count <- connectionRepository
-        .updateWithConnectionResponse(recordId, response, ProtocolState.ConnectionResponsePending)
+        .updateWithConnectionResponse(recordId, response, ProtocolState.ConnectionResponsePending, maxRetries)
         .mapError(RepositoryError.apply)
       _ <- count match
         case 1 => ZIO.succeed(())
@@ -175,7 +180,7 @@ private class ConnectionServiceImpl(
     updateConnectionProtocolState(
       recordId,
       ProtocolState.ConnectionResponsePending,
-      ProtocolState.ConnectionResponseSent
+      ProtocolState.ConnectionResponseSent,
     )
 
   override def receiveConnectionResponse(
@@ -188,7 +193,7 @@ private class ConnectionServiceImpl(
         ProtocolState.ConnectionRequestSent
       )
       _ <- connectionRepository
-        .updateWithConnectionResponse(record.id, response, ProtocolState.ConnectionResponseReceived)
+        .updateWithConnectionResponse(record.id, response, ProtocolState.ConnectionResponseReceived, maxRetries)
         .flatMap {
           case 1 => ZIO.succeed(())
           case n => ZIO.fail(UnexpectedException(s"Invalid row count result: $n"))
@@ -232,11 +237,11 @@ private class ConnectionServiceImpl(
   private[this] def updateConnectionProtocolState(
       recordId: UUID,
       from: ProtocolState,
-      to: ProtocolState
+      to: ProtocolState,
   ): IO[ConnectionServiceError, Option[ConnectionRecord]] = {
     for {
       _ <- connectionRepository
-        .updateConnectionProtocolState(recordId, from, to)
+        .updateConnectionProtocolState(recordId, from, to, maxRetries)
         .flatMap {
           case 1 => ZIO.succeed(())
           case n => ZIO.fail(UnexpectedException(s"Invalid row count result: $n"))