+++ pollux/lib/vc-jwt/src/test/scala/io/iohk/atala/pollux/vc/jwt/MySuite.scala
@@ -1,9 +1,0 @@
-package io.iohk.atala.pollux.vc.jwt
-
-class MySuite extends munit.FunSuite {
-  test("example test that succeeds") {
-    val obtained = 42
-    val expected = 42
-    assertEquals(obtained, expected)
-  }
-}